fix: stop -server.toml being reset to defaults on every load
ModConfigSpec.define(path, default) derives its validator from the
default value's runtime class, accepting a loaded value only if it is an
instance of that exact boxed type. The Forge Config API Port TOML parser
reads a whole number like `62` as Integer, but the Y-level / threshold
settings were defined with Long defaults (62L) — so every load the value
was judged "incorrect", reset to the default, and the file rewritten with
a .bak. The file watcher then re-read that rewrite and looped forever, so
edits to those values never stuck.

- Type whole-number settings as int (depth/nether startY/endY,
  softCap.threshold) to match what the parser produces; this also removes
  the five (int) casts in HardnessCalculator and makes HardnessSettings
  and its stream codec honest int.
- Keep decimal settings lenient (accept `2` as well as `2.0`) via an
  `o -> o instanceof Number` validator.
- Use defineInRange(0.0, 1.0) for the two genuinely bounded fractions
  (softCap.multiplier, scope.exemptMultiplier).

+++ common/src/gametest/java/dev/gimme/mininghardness/gametest/MiningHardnessGameTests.java
@@ -137,13 +137,13 @@ public final class MiningHardnessGameTests {
     /**
      * The streamed {@link HardnessSettings} survives a network encode→decode unchanged. Guards the hand-written
      * {@code STREAM_CODEC} against silent field drift — in particular a reorder of two same-typed fields (there are five
-     * {@code long}s and several {@code double}s), which the compiler can't catch but which would swap their values on
+     * {@code int}s and several {@code double}s), which the compiler can't catch but which would swap their values on
      * the wire and desync client prediction. Every field gets a distinct value, so any such swap fails the round-trip;
-     * the leftover-bytes check additionally catches a read/write width mismatch (e.g. {@code readInt} vs {@code writeLong}).
+     * the leftover-bytes check additionally catches a read/write width mismatch (e.g. {@code readInt} vs {@code writeDouble}).
      */
     public static void configSettingsSurviveNetworkRoundTrip(GameTestHelper helper) {
         HardnessSettings original = new HardnessSettings(
-                11L, -22L, 3.5, 4.5, 5.5, 66L, -77L, 88L, 9.5, 10.5, 11.5, "stone|deepslate", ".*_ore", 12.5);
+                11, -22, 3.5, 4.5, 5.5, 66, -77, 88, 9.5, 10.5, 11.5, "stone|deepslate", ".*_ore", 12.5);
 
         FriendlyByteBuf buf = new FriendlyByteBuf(Unpooled.buffer());
         HardnessSettings.STREAM_CODEC.encode(buf, original);
@@ -221,7 +221,7 @@ public final class MiningHardnessGameTests {
 
     /** Pins depth scaling so the block at {@code pos} sits exactly at full depth, with depth alone doubling hardness. */
     private static ConfigTestSupport.Scope pinFullDepthAt(BlockPos pos) {
-        long y = pos.getY();
+        int y = pos.getY();
         var startY = ConfigTestSupport.override(ConfigTestSupport.DEPTH_START_Y, y + 1);
         var endY = ConfigTestSupport.override(ConfigTestSupport.DEPTH_END_Y, y);
         var bonus = ConfigTestSupport.override(ConfigTestSupport.DEPTH_MULTIPLIER_BONUS, 1.0);
